requirements.txt
django
gunicorn
psycopg2
django-summernote
django-bootstrap-datepicker-plus
django-taggit
django-multiupload
django-imagekit
pillow
pilkit
dj-database-url